+++ cocos/core/pipeline/index.jsb.ts
@@ -26,9 +26,10 @@
 declare const nr: any;
 
 import { getPhaseID } from './pass-phase';
-import { js, setClassName } from '../../core/utils/js';
+import { setClassName } from '../../core/utils/js';
+import { PipelineEventType } from './pipeline-event';
 import * as pipeline from './define';
-import { DebugView } from './debug-view';
+export { pipeline };
 
 nr.getPhaseID = getPhaseID;
 
@@ -48,30 +49,19 @@ export const LightingStage = nr.LightingStage;
 export const PostProcessStage = nr.PostProcessStage;
 export const GbufferStage = nr.GbufferStage;
 export const BloomStage = nr.BloomStage;
-
-export { pipeline };
 export { GeometryRenderer } from './geometry-renderer';
 export { PipelineEventType } from './pipeline-event';
-export { DebugView };
 
 let getOrCreatePipelineState = nr.PipelineStateManager.getOrCreatePipelineState;
 nr.PipelineStateManager.getOrCreatePipelineState = function(device, pass, shader, renderPass, ia) {
     return getOrCreatePipelineState(pass, shader, renderPass, ia); //cjh TODO: remove hacking. c++ API doesn't access device argument.
 };
 
-// RenderPipeline
-const renderPipelineProto = RenderPipeline.prototype;
-js.get(renderPipelineProto, 'debugView', function () {
-    return this._debugView;
-})
-
 // ForwardPipeline
 const forwardPipelineProto = ForwardPipeline.prototype;
 forwardPipelineProto._ctor = function() {
     this._tag = 0;
     this._flows = [];
-    this._debugView = new DebugView();
-    this.setDebugViewConfig(this._debugView._nativeConfig);
 };
 
 forwardPipelineProto.init = function () {
@@ -173,8 +163,6 @@ deferredPipelineProto._ctor = function() {
     this._flows = [];
     this.renderTextures = [];
     this.materials = [];
-    this._debugView = new DebugView();
-    this.setDebugViewConfig(this._debugView._nativeConfig);
 }
 
 const oldDeferredOnLoaded = deferredPipelineProto.onLoaded;
