+++ cocos/core/pipeline/index.ts
@@ -53,8 +53,7 @@ export { InstancedBuffer } from './instanced-buffer';
 export { PipelineStateManager } from './pipeline-state-manager';
 
 export { PipelineEventProcessor, PipelineEventType } from './pipeline-event';
-
-export { DebugView, DebugViewSingleType, DebugViewCompositeType } from './debug-view';
+export { DebugView } from './debug-view';
 
 export function createDefaultPipeline () {
     const rppl = new ForwardPipeline();
