+++ cocos/core/root.jsb.ts
@@ -5,6 +5,7 @@ import legacyCC from '../../predefine';
 import { DataPoolManager } from '../3d/skeletal-animation/data-pool-manager';
 import { Device } from './gfx';
 import { registerRebuildLayoutGraph } from './pipeline/custom/index.jsb';
+import { DebugView } from './pipeline/debug-view';
 
 declare const nr: any;
 declare const jsb: any;
@@ -52,6 +53,14 @@ Object.defineProperty(rootProto, 'pipelineEvent', {
     }
 });
 
+/*Object.defineProperty(rootProto, 'debugView', {
+    configurable: true,
+    enumerable: true,
+    get () {
+        return this._debugView;
+    }
+});*/
+
 class DummyPipelineEvent {
     on (type: any, callback: any, target?: any, once?: boolean) {}
     once (type: any, callback: any, target?: any) {}
@@ -69,6 +78,10 @@ rootProto._ctor = function (device: Device) {
     this._lightPools = new Map();
     this._batcher = null;
     this._pipelineEvent = new DummyPipelineEvent();
+    // this._debugView = new DebugView();
+    // this.setDebugViewConfig(this._debugView._nativeConfig);
+    this._debugView = new DebugView();
+    this.setDebugViewConfig(this._debugView._nativeConfig);
     this._registerListeners();
 };
